my组件取消ã.catch和header的注销修复

+++ src/pages/My/template.vue
@@ -69,15 +69,18 @@
         })
       },
       async onDelete(blogId) {
-        await this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
-          confirmButtonText: '确定',
-          cancelButtonText: '取消',
-          type: 'warning'
-        })
-        await blog.deleteBlog({blogId})
-        this.$message.success('删除成功')
-        this.blogs = this.blogs.filter(blog => blog.id !== blogId)
+          await this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText: '取消',
+            type: 'warning'
+          }) .then(async () => {
+            await blog.deleteBlog({blogId})
+            this.$message.success('删除成功')
+            this.blogs = this.blogs.filter(blog => blog.id !== blogId)
+          }).catch(() => {})
       },
+
+
       splitDate(dataStr) {
         let dateObj = typeof dataStr === 'object' ? dataStr : new Date(dataStr)
         return {
